rework: simpler documentation
Updated and Added documents that were located somewhere else.
This will make it easier to integrate the documentation and
update it accordingly. We are also using several features
provided by GitHub, to make documentation discovery easier.

Also, I added the new commit guideline and started to work
on the GitHub page via mkdocs.